chore: split type tests between scripts and contracts
